pyproject.toml
[project]
name = "atlas"
version = "0.1.0"
description = "Personal AI agent — model-agnostic, extensible, runs locally"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "python-dotenv>=1.0.0",
    "python-telegram-bot>=21.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ddgs>=9.11.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
